# apps/api/src/generated/prisma/browser.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
 * This file should be your main import to use Prisma-related types and utilities in a browser. 
 * Use it to get access to models, enums, and input types.
 * 
 * This file does not contain a `PrismaClient` class, nor several other helpers that are intended as server-side only.
 * See `client.ts` for the standard, server-side entry point.
 *
 * 🟢 You can import this file directly.
 */

import * as Prisma from './internal/prismaNamespaceBrowser.js'
export { Prisma }
export * as $Enums from './enums.js'
export * from './enums.js';
/**
 * Model Profile
 * 
 */
export type Profile = Prisma.ProfileModel
/**
 * Model Household
 * 
 */
export type Household = Prisma.HouseholdModel
/**
 * Model HouseholdMember
 * 
 */
export type HouseholdMember = Prisma.HouseholdMemberModel
/**
 * Model HouseholdInvitation
 * 
 */
export type HouseholdInvitation = Prisma.HouseholdInvitationModel
/**
 * Model Room
 * 
 */
export type Room = Prisma.RoomModel
/**
 * Model TaskTemplate
 * 
 */
export type TaskTemplate = Prisma.TaskTemplateModel
/**
 * Model Task
 * 
 */
export type Task = Prisma.TaskModel
/**
 * Model TaskCompletion
 * 
 */
export type TaskCompletion = Prisma.TaskCompletionModel
/**
 * Model DailyPlan
 * 
 */
export type DailyPlan = Prisma.DailyPlanModel
/**
 * Model DailyPlanItem
 * 
 */
export type DailyPlanItem = Prisma.DailyPlanItemModel
/**
 * Model Device
 * 
 */
export type Device = Prisma.DeviceModel
/**
 * Model NotificationPreference
 * 
 */
export type NotificationPreference = Prisma.NotificationPreferenceModel