# apps/api/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const HouseholdRole = {
  owner: 'owner',
  member: 'member'
} as const

export type HouseholdRole = (typeof HouseholdRole)[keyof typeof HouseholdRole]


export const RoomType = {
  kitchen: 'kitchen',
  bathroom: 'bathroom',
  bedroom: 'bedroom',
  living_room: 'living_room',
  hallway: 'hallway',
  office: 'office',
  other: 'other'
} as const

export type RoomType = (typeof RoomType)[keyof typeof RoomType]


export const Effort = {
  low: 'low',
  normal: 'normal',
  high: 'high'
} as const

export type Effort = (typeof Effort)[keyof typeof Effort]


export const TaskPriority = {
  low: 'low',
  normal: 'normal',
  high: 'high'
} as const

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority]


export const AssignmentMode = {
  unassigned: 'unassigned',
  fixed: 'fixed',
  rotation: 'rotation'
} as const

export type AssignmentMode = (typeof AssignmentMode)[keyof typeof AssignmentMode]


export const PlanStatus = {
  draft: 'draft',
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled'
} as const

export type PlanStatus = (typeof PlanStatus)[keyof typeof PlanStatus]


export const PlanItemStatus = {
  planned: 'planned',
  completed: 'completed',
  skipped: 'skipped',
  deferred: 'deferred'
} as const

export type PlanItemStatus = (typeof PlanItemStatus)[keyof typeof PlanItemStatus]


export const SelectionReason = {
  overdue: 'overdue',
  due: 'due',
  priority: 'priority',
  manual: 'manual'
} as const

export type SelectionReason = (typeof SelectionReason)[keyof typeof SelectionReason]


export const DevicePlatform = {
  ios: 'ios',
  android: 'android'
} as const

export type DevicePlatform = (typeof DevicePlatform)[keyof typeof DevicePlatform]