# outputs/openapi.yaml
openapi: 3.1.0
info:
  title: Home Cleaning Planner API
  version: 0.1.0
  description: REST API contract for the cleaning planner MVP.
servers:
  - url: https://api.example.com/v1
    description: Production placeholder
  - url: http://localhost:3000/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /task-templates:
    get:
      tags: [Templates]
      operationId: listTaskTemplates
      summary: List localized task templates
      parameters:
        - name: roomType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RoomType"
        - name: locale
          in: query
          required: false
          schema:
            type: string
            default: uk-UA
      responses:
        "200":
          description: Task templates
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /rooms/{roomId}/tasks:
    get:
      tags: [Tasks]
      operationId: listRoomTasks
      summary: List tasks in a room
      parameters:
        - $ref: "#/components/parameters/RoomId"
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: dueState
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/DueState"
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskSummary"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Tasks]
      operationId: createTask
      summary: Create a task manually or from a template
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /tasks/{taskId}:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      tags: [Tasks]
      operationId: getTask
      summary: Get task details
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Tasks]
      operationId: updateTask
      summary: Update or restore a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Tasks]
      operationId: archiveTask
      summary: Archive a task
      parameters:
        - $ref: "#/components/parameters/ExpectedVersion"
      responses:
        "204":
          description: Task archived
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /tasks/{taskId}/completions:
    get:
      tags: [Tasks]
      operationId: listTaskCompletions
      summary: List task completion history
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Completion history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompletionPage"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks/{taskId}/complete:
    post:
      tags: [Tasks]
      operationId: completeTask
      summary: Complete a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteTaskRequest"
      responses:
        "201":
          description: Completion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteTaskResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /task-completions/{completionId}/undo:
    post:
      tags: [Tasks]
      operationId: undoTaskCompletion
      summary: Undo a task completion without deleting history
      parameters:
        - $ref: "#/components/parameters/CompletionId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Completion undone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteTaskResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /households/{householdId}/due-tasks:
    get:
      tags: [Tasks]
      operationId: listDueTasks
      summary: List due, upcoming, and overdue household tasks
      parameters:
        - $ref: "#/components/parameters/HouseholdId"
        - name: localDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: timezone
          in: query
          required: true
          schema:
            type: string
            example: Europe/Madrid
        - name: assignee
          in: query
          required: false
          schema:
            type: string
            enum: [me, anyone]
            default: me
      responses:
        "200":
          description: Due task groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DueTasksResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token.
  parameters:
    HouseholdId:
      name: householdId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CompletionId:
      name: completionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
    ExpectedVersion:
      name: expectedVersion
      in: query
      required: true
      schema:
        type: integer
        minimum: 1
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 30
  responses:
    Unauthorized:
      description: Authentication is required or the token is invalid
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: Resource not found within the authorization boundary
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Domain or resource version conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds before retrying.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    RoomType:
      type: string
      enum: [kitchen, bathroom, bedroom, living_room, hallway, office, other]
...
    Effort:
      type: string
      enum: [low, normal, high]
    Priority:
      type: string
      enum: [low, normal, high]
    DueState:
      type: string
      enum: [not_due, upcoming, due, overdue]
    AssignmentMode:
      type: string
      enum: [unassigned, fixed, rotation]
    IntervalRecurrence:
      type: object
      additionalProperties: false
      required: [type, every, unit]
      properties:
        type:
          type: string
          const: interval
        every:
          type: integer
          minimum: 1
          maximum: 365
        unit:
          type: string
          enum: [day, week]
    WeeklyRecurrence:
      type: object
      additionalProperties: false
      required: [type, daysOfWeek]
      properties:
        type:
          type: string
          const: weekly
        daysOfWeek:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: integer
            minimum: 1
            maximum: 7
    MonthlyRecurrence:
      type: object
      additionalProperties: false
      required: [type, dayOfMonth]
      properties:
        type:
          type: string
          const: monthly
        dayOfMonth:
          type: integer
          minimum: 1
          maximum: 31
    ManualRecurrence:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          type: string
          const: manual
    Recurrence:
      oneOf:
        - $ref: "#/components/schemas/IntervalRecurrence"
        - $ref: "#/components/schemas/WeeklyRecurrence"
        - $ref: "#/components/schemas/MonthlyRecurrence"
        - $ref: "#/components/schemas/ManualRecurrence"
      discriminator:
        propertyName: type
    TaskTemplate:
      type: object
      required: [id, roomType, locale, title, estimatedMinutes, effort, recurrence]
      properties:
        id:
          type: string
          format: uuid
        roomType:
          $ref: "#/components/schemas/RoomType"
        locale:
          type: string
        title:
          type: string
        estimatedMinutes:
          type: integer
          minimum: 1
        effort:
          $ref: "#/components/schemas/Effort"
        recurrence:
          $ref: "#/components/schemas/Recurrence"
    TaskSummary:
      type: object
      required:
        - id
        - householdId
        - roomId
        - title
        - estimatedMinutes
        - effort
        - priority
        - recurrence
        - assignmentMode
        - rotationMemberIds
        - dueState
        - archived
        - version
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        title:
          type: string
        estimatedMinutes:
          type: integer
        effort:
          $ref: "#/components/schemas/Effort"
        priority:
          $ref: "#/components/schemas/Priority"
        recurrence:
          $ref: "#/components/schemas/Recurrence"
        assignmentMode:
          $ref: "#/components/schemas/AssignmentMode"
        fixedAssigneeId:
          type: [string, "null"]
          format: uuid
        rotationMemberIds:
          type: array
          items:
            type: string
            format: uuid
        dueState:
          $ref: "#/components/schemas/DueState"
        lastCompletedAt:
          type: [string, "null"]
          format: date-time
        nextDueOn:
          type: [string, "null"]
          format: date
        archived:
          type: boolean
        version:
          type: integer
          minimum: 1
    Task:
      allOf:
        - $ref: "#/components/schemas/TaskSummary"
        - type: object
          required: [createdBy, createdAt, updatedAt]
          properties:
            templateId:
              type: [string, "null"]
              format: uuid
            notes:
              type: [string, "null"]
            rotationIndex:
              type: integer
              minimum: 0
            createdBy:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required: [title, estimatedMinutes, effort, priority, recurrence, assignmentMode]
      properties:
        templateId:
          type: [string, "null"]
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 120
        notes:
          type: [string, "null"]
          maxLength: 2000
        estimatedMinutes:
          type: integer
          minimum: 1
          maximum: 480
        effort:
          $ref: "#/components/schemas/Effort"
        priority:
          $ref: "#/components/schemas/Priority"
        recurrence:
          $ref: "#/components/schemas/Recurrence"
        assignmentMode:
          $ref: "#/components/schemas/AssignmentMode"
        fixedAssigneeId:
          type: [string, "null"]
          format: uuid
        rotationMemberIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
    UpdateTaskRequest:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
        title:
          type: string
          minLength: 1
          maxLength: 120
        notes:
          type: [string, "null"]
          maxLength: 2000
        estimatedMinutes:
          type: integer
          minimum: 1
          maximum: 480
        effort:
          $ref: "#/components/schemas/Effort"
        priority:
          $ref: "#/components/schemas/Priority"
        recurrence:
          $ref: "#/components/schemas/Recurrence"
        assignmentMode:
          $ref: "#/components/schemas/AssignmentMode"
        fixedAssigneeId:
          type: [string, "null"]
          format: uuid
        rotationMemberIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        archived:
          type: boolean
    CompleteTaskRequest:
      type: object
      additionalProperties: false
      properties:
        completedAt:
          type: string
          format: date-time
        actualMinutes:
          type: [integer, "null"]
          minimum: 1
          maximum: 1440
        note:
          type: [string, "null"]
          maxLength: 1000
    TaskCompletion:
      type: object
      required: [id, taskId, completedBy, completedAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        completedBy:
          type: string
          format: uuid
        completedAt:
          type: string
          format: date-time
        actualMinutes:
          type: [integer, "null"]
        note:
          type: [string, "null"]
        undoneAt:
          type: [string, "null"]
          format: date-time
        undoneBy:
          type: [string, "null"]
          format: uuid
        createdAt:
          type: string
          format: date-time
    CompleteTaskResponse:
      type: object
      required: [completion, task]
      properties:
        completion:
          $ref: "#/components/schemas/TaskCompletion"
        task:
          $ref: "#/components/schemas/TaskSummary"
    TaskCompletionPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskCompletion"
        nextCursor:
          type: [string, "null"]
    DueTasksResponse:
      type: object
      required: [overdue, due, upcoming]
      properties:
        overdue:
          type: array
          items:
            $ref: "#/components/schemas/TaskSummary"
        due:
          type: array
          items:
            $ref: "#/components/schemas/TaskSummary"
        upcoming:
          type: array
          items:
            $ref: "#/components/schemas/TaskSummary"
...
    Problem:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
        detail:
          type: string
        traceId:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
        currentResource:
          type: [object, "null"]
          additionalProperties: true